altered the ft_itoa_base to accept long long. altered dioux conversions to take in long longs

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: piaandersin <piaandersin@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/17 14:11:54 by pandersi          #+#    #+#             */
-/*   Updated: 2020/03/25 16:38:44 by piaandersin      ###   ########.fr       */
+/*   Updated: 2020/03/26 13:43:14 by piaandersin      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@
 # include <unistd.h>
 # include <stdlib.h>
 # include <string.h>
+# include <limits.h>
 
 typedef struct		s_list
 {
@@ -70,7 +71,7 @@ char				*ft_strncat(char *restrict s1, const char *restrict s2,
 int					ft_strncmp(const char *s1, const char *s2, size_t n);
 char				*ft_strncpy(char *dst, const char *src, size_t len);
 char				*ft_itoa(int n);
-char				*ft_itoa_base(int value, int base);
+char				*ft_itoa_base(long long int value, int base);
 char				*ft_itoa_base_unsigned(size_t value, int base);
 char				*ft_float_to_a(long double nb, unsigned int precision);
 size_t				ft_count_nbr_length(long long n);
